@@ -28,8 +28,10 @@ set(ABACUS_TEST_DIR ${CMAKE_CURRENT_SOURCE_DIR}/tests)
 set(ABACUS_BIN_PATH ${CMAKE_CURRENT_BINARY_DIR}/${ABACUS_BIN_NAME})
 include_directories(${ABACUS_SOURCE_DIR})
 add_executable(${ABACUS_BIN_NAME} source/main.cpp)
+set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
 add_compile_options(-O2 -g)
-SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-write-strings" )
+add_compile_options(-Wno-write-strings)
 
 find_package(Cereal REQUIRED)
 include_directories(${Cereal_INCLUDE_DIR})
@@ -244,10 +246,9 @@ add_compile_definitions(
 )
 
 IF (BUILD_TESTING)
+  set(CMAKE_CXX_STANDARD 14) # Required in orbital
   include(CTest)
   enable_testing()
-  find_package(Threads)
-  set(CMAKE_CXX_STANDARD_REQUIRED ON)
   find_package(GTest HINTS /usr/local/lib/ ${GTEST_DIR})
   if(NOT ${GTest_FOUND})
     include(FetchContent)
