@@ -28,7 +28,11 @@ namespace ModuleESolver
 
         // calculate electron density from a specific Hamiltonian
         virtual void hamilt2density(const int istep, const int iter, const double ethr);
-        // get
+
+        // calculate electron states from a specific Hamiltonian
+        virtual void hamilt2estates(const double ethr){};
+
+        // get current step of Ionic simulation
         virtual int getniter() override;
 
     protected:
