@@ -38,9 +38,15 @@ template <typename T> class LocalMatrix
 template <typename T, typename T1> class HamiltLCAO : public Hamilt
 {
   public:
-    HamiltLCAO(Gint_Gamma* GG_in, Gint_k* GK_in, LCAO_gen_fixedH* genH_in, LCAO_Matrix* LM_in)
+    HamiltLCAO(Gint_Gamma* GG_in, LCAO_gen_fixedH* genH_in, LCAO_Matrix* LM_in)
     {
         this->GG = GG_in;
+        this->genH = genH_in;
+        this->LM = LM_in;
+        this->classname = "HamiltLCAO";
+    }
+    HamiltLCAO(Gint_k* GK_in, LCAO_gen_fixedH* genH_in, LCAO_Matrix* LM_in)
+    {
         this->GK = GK_in;
         this->genH = genH_in;
         this->LM = LM_in;
