@@ -26,11 +26,25 @@ class HSolver
         Input &in )=0;*/
 
     // solve Hamiltonian to electronic density in ElecState
-    virtual void solve(hamilt::Hamilt* phm, psi::Psi<std::complex<double>>& ppsi, elecstate::ElecState* pes)
+    virtual void solve
+    (
+        hamilt::Hamilt* phm, 
+        psi::Psi<std::complex<double>>& ppsi, 
+        elecstate::ElecState* pes, 
+        const std::string method, 
+        const bool skip_charge=false
+    )
     {
         return;
     }
-    virtual void solve(hamilt::Hamilt* phm, psi::Psi<double>& ppsi, elecstate::ElecState* pes)
+    virtual void solve
+    (
+        hamilt::Hamilt* phm, 
+        psi::Psi<double>& ppsi, 
+        elecstate::ElecState* pes, 
+        const std::string method, 
+        const bool skip_charge=false
+    )
     {
         return;
     }
