@@ -90,12 +90,20 @@ class Psi
     // choose k-point index , then Psi(iband, ibasis) can reach Psi(ik, iband, ibasis)
     void fix_k(const int ik) const
     {
-        assert(ik>=0 && ik<this->nk);
+        assert(ik>=0);
         this->current_k = ik;
         if(this->ngk!=nullptr&&GlobalV::NSPIN!=4) this->current_nbasis = this->ngk[ik];
         else this->current_nbasis = this->nbasis;
         this->current_b = 0;
-        this->psi_current = const_cast<T*>(&(this->psi[ik * this->nbands * this->nbasis]));
+        if( ik >= this->nk)
+        {
+            // mem_saver case
+            this->psi_current = const_cast<T*>(&(this->psi[0]));
+        }
+        else
+        {
+            this->psi_current = const_cast<T*>(&(this->psi[ik * this->nbands * this->nbasis]));
+        }
         return;
     }
 
