@@ -124,8 +124,9 @@ void Ions::opt_ions_pw(ModuleESolver::ESolver *p_esolver)
         }
         else if(GlobalV::CALCULATION=="nscf")
         {
-            elec.non_self_consistent(istep-1);
-			eiter = elec.iter;
+			p_esolver->nscf();
+            //elec.non_self_consistent(istep-1);
+			eiter = p_esolver->getniter();
         }
 
 		if(GlobalC::pot.out_pot == 2)
@@ -185,11 +186,6 @@ void Ions::opt_ions_pw(ModuleESolver::ESolver *p_esolver)
 		std::cout << " ION DYNAMICS FINISHED :)" << std::endl;
 	}
 
-	if(GlobalC::wf.out_wfc_r == 1)				// Peize Lin add 2021.11.21
-	{
-		Write_Wfc_Realspace::write_wfc_realspace_1(GlobalC::wf.psi[0], "wfc_realspace", true);
-	}	
-
 	ModuleBase::timer::tick("Ions","opt_ions_pw");
     return;
 }
@@ -236,7 +232,7 @@ bool Ions::after_scf(ModuleESolver::ESolver *p_esolver, const int &istep, int &f
 		//do cell relax calculation and generate next structure
 		bool converged = 0;
 		converged = this->do_cellrelax(stress_step, stress, GlobalC::en.etot);
-		if(!converged) this->reset_after_cellrelax(force_step, stress_step);
+		if(!converged) this->reset_after_cellrelax(force_step, stress_step, p_esolver);
 		return converged;
 	}
 
@@ -344,14 +340,14 @@ void Ions::reset_after_relax(const int& istep)
 	GlobalV::ofs_running << " Setup the new wave functions?" << std::endl;
 	//GlobalC::wf.wfcinit();
 }
-void Ions::reset_after_cellrelax(int& f_step, int& s_step)
+void Ions::reset_after_cellrelax(int& f_step, int& s_step, ModuleESolver::ESolver *p_esolver)
 {
 	ModuleBase::TITLE("Ions","reset_after_cellrelax");
-	Variable_Cell::init_after_vc();
+	Variable_Cell::init_after_vc(p_esolver);
 	GlobalC::pot.init_pot(s_step, GlobalC::pw.strucFac); //LiuXh add 20180619
 
 	GlobalV::ofs_running << " Setup the new wave functions?" << std::endl; //LiuXh add 20180619
-	GlobalC::wf.wfcinit(); //LiuXh add 20180619
+	GlobalC::wf.wfcinit(p_esolver->psi); //LiuXh add 20180619
 	f_step = 1;
 	++s_step;
 }
