@@ -19,8 +19,7 @@ Forces::Forces()
 Forces::~Forces() {}
 
 #include "../module_base/mathzone.h"
-
-void Forces::init(ModuleBase::matrix& force)
+void Forces::init(ModuleBase::matrix& force, const psi::Psi<complex<double>>* psi_in)
 {
 	ModuleBase::TITLE("Forces", "init");
 	this->nat = GlobalC::ucell.nat;
@@ -33,7 +32,7 @@ void Forces::init(ModuleBase::matrix& force)
 	ModuleBase::matrix forcescc(nat, 3);
     this->cal_force_loc(forcelc, GlobalC::rhopw);
     this->cal_force_ew(forceion, GlobalC::rhopw);
-    this->cal_force_nl(forcenl);
+    this->cal_force_nl(forcenl, psi_in);
 	this->cal_force_cc(forcecc, GlobalC::rhopw);
 	this->cal_force_scc(forcescc, GlobalC::rhopw);
 
@@ -660,7 +659,7 @@ void Forces::cal_force_cc(ModuleBase::matrix& forcecc, ModulePW::PW_Basis* rho_b
 
 #include "../module_base/complexarray.h"
 #include "../module_base/complexmatrix.h"
-void Forces::cal_force_nl(ModuleBase::matrix& forcenl)
+void Forces::cal_force_nl(ModuleBase::matrix& forcenl, const psi::Psi<complex<double>>* psi_in)
 {
 	ModuleBase::TITLE("Forces","cal_force_nl");
 	ModuleBase::timer::tick("Forces","cal_force_nl");
@@ -695,7 +694,7 @@ void Forces::cal_force_nl(ModuleBase::matrix& forcenl)
         {
             for (int i=0;i<nkb;i++)
             {
-                const std::complex<double>* ppsi = &(GlobalC::wf.psi[0](ik, ib, 0));
+                const std::complex<double>* ppsi = &(psi_in[0](ik, ib, 0));
                 const std::complex<double>* pvkb = &(GlobalC::ppcell.vkb(i, 0));
                 for (int ig=0; ig<nbasis; ig++)
                 {
@@ -737,7 +736,7 @@ void Forces::cal_force_nl(ModuleBase::matrix& forcenl)
                 if(GlobalC::wf.wg(ik, ib) < ModuleBase::threshold_wg) continue;
                 for (int i=0; i<nkb; i++)
                 {
-                    const std::complex<double>* ppsi = &(GlobalC::wf.psi[0](ik, ib, 0));
+                    const std::complex<double>* ppsi = &(psi_in[0](ik, ib, 0));
                     const std::complex<double>* pvkb1 = &(vkb1(i, 0));
                     for (int ig=0; ig<nbasis; ig++)
                     {
