@@ -96,7 +96,7 @@ void Run_MD_PW::md_ions_pw(ModuleESolver::ESolver *p_esolver)
 
             if(cellchange)
             {
-                Variable_Cell::init_after_vc();
+                Variable_Cell::init_after_vc(p_esolver);
             }
 
             // reset local potential and initial wave function
