@@ -8,10 +8,9 @@ class Stress_PW: public Stress_Func
 	public :
 	
 	Stress_PW (){};
-	~Stress_PW (){};
 
 	//calculate the stress in PW basis
-	void cal_stress(ModuleBase::matrix& smearing_sigmatot);
+	void cal_stress(ModuleBase::matrix& smearing_sigmatot, const psi::Psi<complex<double>>* psi_in=nullptr);
 
 	protected :
 
